Add an optimization in advance_minute
This optimization is for the special case where self.minutes
has exactly one element. Instead of advancing one minute per
iteration, make a jump from the current minute to the target
minute.

In a benchmark over 5000+ different cron expressions, this
change improves iterator's performance by about 2x. Given that
the optimization is isolated and relatively straightforward,
I think it is worth it!