Change day-of-month and day-of-week handling to mimic Debian cron
In Debian cron, if either the day-of-month or the day-of-week field
starts with a star, then there is an "AND" relationship between them.
Otherwise it's "OR".

This commit updates cronsim to follow the same logic.

Example expression: 1 1 1-7 * */7

Here, the day-of-month is "1-7", meaning dates from 1 to 7. The
day-of-the-week field is "*/7", meaning "every 7th weekday, starting
from Sunday". So, effectively, "every Sunday". Since the day-of-week
field starts with a star, there's an AND relationship between
the day-of-month and the day-of-week fields, and the expression
translates to "at 01:01 of the first Sunday of every month".